--- testdata/zhJoinLines.text
@@ -1,8 +1,0 @@
-# 标题
-
-第一
-行文字。
-
-第
-二
-行文字。
